move logic to cli interface
