+++ index.js
@@ -6,7 +6,6 @@ const chokidar = require('chokidar')
 const arrify = require('arrify')
 const { Deque } = require('@blakeembrey/deque')
 const { supportsColor } = require('supports-color')
-const { readFileSync, existsSync } = require('fs')
 
 const ECHO_JS_PATH = resolve(__dirname, 'echo.js')
 const ECHO_CMD = `${quote(process.execPath)} ${quote(ECHO_JS_PATH)}`
@@ -118,15 +117,11 @@ function onchange (match, command, rawArgs, opts = {}) {
   if (!command && !outpipe) {
     throw new TypeError('Expected "command" and/or "outpipe" to be specified')
   }
-
-  const ignored = getIgnoreMergedFromIgnoreFile(opts.exclude, opts.ignorePath)
-
-  console.log(ignored)
   
   // Create the "watcher" instance for file system changes.
   const watcher = chokidar.watch(matches, {
     cwd: cwd,
-    ignored,
+    ignored: opts.exclude || [],
     ignoreInitial: opts.add !== true,
     usePolling: opts.poll === true || typeof opts.poll === 'number',
     interval: typeof opts.poll === 'number' ? opts.poll : undefined,
@@ -196,21 +191,6 @@ function onchange (match, command, rawArgs, opts = {}) {
   })
 
   watcher.on('error', (error) => log(`watcher error: ${error}`))
-}
-
-function getIgnoreMergedFromIgnoreFile(exclude = [], ignorePath) {
-    if(ignorePath && existsSync(ignorePath)) {
-      const ignoreFileString = readFileSync(ignorePath).toString('utf-8')
-      const ignoreFileArray = ignoreFileString.replace(/^#[^\r\n]+\r?\n/gm, '').split(/\r?\n/)
-      
-      const ignoredSet = new Set([...ignoreFileArray, ...exclude])
-      
-      ignoredSet.delete('')
-  
-      return [...ignoredSet]
-    }
-
-    return exclude
 }
 
 // Double mustache template generator.
